Suppress various harmless warnings in MinGW build with -Wconversion.
No real changes, simply add casts to make the implicit conversions that g++
warns about when using -Wconversion explicit.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@67964 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/slider.cpp
@@ -602,7 +602,7 @@ wxSize wxSlider::DoGetBestSize() const
             if ( HasFlag(wxSL_MIN_MAX_LABELS) )
                 size.y += labelSize;
             if ( HasFlag(wxSL_VALUE_LABEL) )
-                size.y += labelSize*2.75;
+                size.y += static_cast<int>(labelSize*2.75);
         }
     }
